requirements.txt
pygame-ce==2.2.1
ujson==5.7.0
pygame_gui==0.6.9
bidict==0.22.1